feat: implement persistent background webhook task queue with worker pool
